refactor: correct shortcut module type ownership
